feat: Order 를 저장하는 기능 추가

+++ src/main/java/cart/dao/OrderProductDao.java
@@ -17,7 +17,7 @@ public class OrderProductDao {
             rs.getLong("id"),
             rs.getLong("order_id"),
             rs.getLong("product_id"),
-            rs.getLong("quantity")
+            rs.getInt("quantity")
     );
 
     private final JdbcTemplate jdbcTemplate;
